+++ flash/text/seq2seq/summarization/model.py
@@ -83,7 +83,7 @@ class SummarizationTask(Seq2SeqTask):
 
     def compute_metrics(self, generated_tokens: torch.Tensor, batch: Dict, prefix: str) -> None:
         tgt_lns = self.tokenize_labels(batch["labels"])
-        result = self.rouge(self._postprocess.uncollate(generated_tokens), tgt_lns)
+        result = self.rouge(self._output_transform.uncollate(generated_tokens), tgt_lns)
         self.log_dict(result, on_step=False, on_epoch=True, prog_bar=True)
 
     @staticmethod