+++ tests/core/data/test_data_pipeline.py
@@ -25,12 +25,13 @@ from torch.utils.data._utils.collate import default_collate
 
 from flash import Trainer
 from flash.core.data.auto_dataset import IterableAutoDataset
-from flash.core.data.batch import _Postprocessor, _Preprocessor
+from flash.core.data.batch import _Preprocessor
 from flash.core.data.data_module import DataModule
 from flash.core.data.data_pipeline import _StageOrchestrator, DataPipeline, DataPipelineState
 from flash.core.data.data_source import DataSource
 from flash.core.data.io.output import Output
-from flash.core.data.process import DefaultPreprocess, Deserializer, Postprocess, Preprocess
+from flash.core.data.io.output_transform import _OutputTransformProcessor, OutputTransform
+from flash.core.data.process import DefaultPreprocess, Deserializer, Preprocess
 from flash.core.data.properties import ProcessState
 from flash.core.data.states import PerBatchTransformOnDevice, ToTensorTransform
 from flash.core.model import Task
@@ -73,23 +74,23 @@ def test_data_pipeline_str():
     data_pipeline = DataPipeline(
         data_source=cast(DataSource, "data_source"),
         preprocess=cast(Preprocess, "preprocess"),
-        postprocess=cast(Postprocess, "postprocess"),
+        output_transform=cast(OutputTransform, "output_transform"),
         output=cast(Output, "output"),
         deserializer=cast(Deserializer, "deserializer"),
     )
 
     expected = "data_source=data_source, deserializer=deserializer, "
-    expected += "preprocess=preprocess, postprocess=postprocess, output=output"
+    expected += "preprocess=preprocess, output_transform=output_transform, output=output"
     assert str(data_pipeline) == (f"DataPipeline({expected})")
 
 
 @pytest.mark.parametrize("use_preprocess", [False, True])
-@pytest.mark.parametrize("use_postprocess", [False, True])
-def test_data_pipeline_init_and_assignement(use_preprocess, use_postprocess, tmpdir):
+@pytest.mark.parametrize("use_output_transform", [False, True])
+def test_data_pipeline_init_and_assignement(use_preprocess, use_output_transform, tmpdir):
     class CustomModel(Task):
-        def __init__(self, postprocess: Optional[Postprocess] = None):
+        def __init__(self, output_transform: Optional[OutputTransform] = None):
             super().__init__(model=torch.nn.Linear(1, 1), loss_fn=torch.nn.MSELoss())
-            self._postprocess = postprocess
+            self._output_transform = output_transform
 
         def train_dataloader(self) -> Any:
             return DataLoader(DummyDataset())
@@ -97,17 +98,17 @@ def test_data_pipeline_init_and_assignement(use_preprocess, use_postprocess, tmp
     class SubPreprocess(DefaultPreprocess):
         pass
 
-    class SubPostprocess(Postprocess):
+    class SubOutputTransform(OutputTransform):
         pass
 
     data_pipeline = DataPipeline(
         preprocess=SubPreprocess() if use_preprocess else None,
-        postprocess=SubPostprocess() if use_postprocess else None,
+        output_transform=SubOutputTransform() if use_output_transform else None,
     )
     assert isinstance(data_pipeline._preprocess_pipeline, SubPreprocess if use_preprocess else DefaultPreprocess)
-    assert isinstance(data_pipeline._postprocess_pipeline, SubPostprocess if use_postprocess else Postprocess)
+    assert isinstance(data_pipeline._output_transform, SubOutputTransform if use_output_transform else OutputTransform)
 
-    model = CustomModel(postprocess=Postprocess())
+    model = CustomModel(output_transform=OutputTransform())
     model.data_pipeline = data_pipeline
     # TODO: the line below should make the same effect but it's not
     # data_pipeline._attach_to_model(model)
@@ -117,10 +118,10 @@ def test_data_pipeline_init_and_assignement(use_preprocess, use_postprocess, tmp
     else:
         assert model._preprocess is None or isinstance(model._preprocess, Preprocess)
 
-    if use_postprocess:
-        assert isinstance(model._postprocess, SubPostprocess)
+    if use_output_transform:
+        assert isinstance(model._output_transform, SubOutputTransform)
     else:
-        assert model._postprocess is None or isinstance(model._postprocess, Postprocess)
+        assert model._output_transform is None or isinstance(model._output_transform, OutputTransform)
 
 
 def test_data_pipeline_is_overriden_and_resolve_function_hierarchy(tmpdir):
@@ -294,9 +295,9 @@ def test_data_pipeline_predict_worker_preprocessor_and_device_preprocessor():
 
 def test_detach_preprocessing_from_model(tmpdir):
     class CustomModel(Task):
-        def __init__(self, postprocess: Optional[Postprocess] = None):
+        def __init__(self, output_transform: Optional[OutputTransform] = None):
             super().__init__(model=torch.nn.Linear(1, 1), loss_fn=torch.nn.MSELoss())
-            self._postprocess = postprocess
+            self._output_transform = output_transform
 
         def train_dataloader(self) -> Any:
             return DataLoader(DummyDataset())
@@ -355,7 +356,7 @@ def test_attaching_datapipeline_to_model(tmpdir):
     class CustomModel(Task):
         def __init__(self):
             super().__init__(model=torch.nn.Linear(1, 1), loss_fn=torch.nn.MSELoss())
-            self._postprocess = Postprocess()
+            self._output_transform = OutputTransform()
 
         def training_step(self, batch: Any, batch_idx: int) -> Any:
             pass
@@ -461,7 +462,7 @@ def test_attaching_datapipeline_to_model(tmpdir):
             assert isinstance(self.predict_step, _StageOrchestrator)
             self._assert_stage_orchestrator_state(self.transfer_batch_to_device._stage_mapping, current_running_stage)
             self._assert_stage_orchestrator_state(
-                self.predict_step._stage_mapping, current_running_stage, cls=_Postprocessor
+                self.predict_step._stage_mapping, current_running_stage, cls=_OutputTransformProcessor
             )
 
         def on_fit_end(self) -> None:
@@ -494,16 +495,20 @@ def test_stage_orchestrator_state_attach_detach(tmpdir):
     _original_predict_step = model.predict_step
 
     class CustomDataPipeline(DataPipeline):
-        def _attach_postprocess_to_model(self, model: "Task", _postprocesssor: _Postprocessor) -> "Task":
-            model.predict_step = self._model_predict_step_wrapper(model.predict_step, _postprocesssor, model)
+        def _attach_output_transform_to_model(
+            self, model: "Task", _output_transform_processor: _OutputTransformProcessor
+        ) -> "Task":
+            model.predict_step = self._model_predict_step_wrapper(
+                model.predict_step, _output_transform_processor, model
+            )
             return model
 
     data_pipeline = CustomDataPipeline(preprocess=preprocess)
-    _postprocesssor = data_pipeline._create_uncollate_postprocessors(RunningStage.PREDICTING)
-    data_pipeline._attach_postprocess_to_model(model, _postprocesssor)
+    _output_transform_processor = data_pipeline._create_output_transform_processor(RunningStage.PREDICTING)
+    data_pipeline._attach_output_transform_to_model(model, _output_transform_processor)
     assert model.predict_step._original == _original_predict_step
-    assert model.predict_step._stage_mapping[RunningStage.PREDICTING] == _postprocesssor
-    data_pipeline._detach_postprocess_from_model(model)
+    assert model.predict_step._stage_mapping[RunningStage.PREDICTING] == _output_transform_processor
+    data_pipeline._detach_output_transform_from_model(model)
     assert model.predict_step == _original_predict_step