+++ tests/text/classification/test_model.py
@@ -23,7 +23,7 @@ from flash.__main__ import main
 from flash.core.data.data_source import DefaultDataKeys
 from flash.core.utilities.imports import _TEXT_AVAILABLE
 from flash.text import TextClassifier
-from flash.text.classification.data import TextClassificationPostprocess, TextClassificationPreprocess
+from flash.text.classification.data import TextClassificationOutputTransform, TextClassificationPreprocess
 from tests.helpers.utils import _SERVE_TESTING, _TEXT_TESTING
 
 # ======== Mock functions ========
@@ -77,9 +77,9 @@ def test_jit(tmpdir):
 @mock.patch("flash._IS_TESTING", True)
 def test_serve():
     model = TextClassifier(2, TEST_BACKBONE)
-    # TODO: Currently only servable once a preprocess and postprocess have been attached
+    # TODO: Currently only servable once a preprocess and output_transform have been attached
     model._preprocess = TextClassificationPreprocess(backbone=TEST_BACKBONE)
-    model._postprocess = TextClassificationPostprocess()
+    model._output_transform = TextClassificationOutputTransform()
     model.eval()
     model.serve()